Reorganise scraper subdirectory
