+++ package.json
@@ -23,7 +23,6 @@
   "license": "MIT",
   "dependencies": {
     "aws-sdk": "^2.36.0",
-    "cheerio": "^0.22.0",
     "isomorphic-fetch": "^2.2.1"
   },
   "devDependencies": {
